don't check return types for abstract methods (#258)
* don't check return types for abstract methods

* Update exp file

+++ test/testdata/compiler/impl_abstract_via_include.rb
@@ -7,7 +7,7 @@ module IFoo
   extend T::Helpers
   interface!
 
-  sig {abstract.returns(Float)}
+  sig {abstract.returns(Integer)}
   def foo; end
 end
